Practise_make_perfect/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
jinja2==3.1.6
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.1.2